bufferedbatch: implement default write function
