+++ bufferedbatch/batch.go
@@ -246,7 +246,33 @@ func (b *Batch) GetValue(key []byte) (value []byte, err error) {
 	return b.originalStore.GetValue(key)
 }
 
-// Write must be overriden by batches that use the BufferedBatch
-func (b *Batch) Write() error {
-	return nil
+// Write implements github.com/stratumn/sdk/store.Batch.Write
+func (b *Batch) Write() (err error) {
+	for _, op := range b.ValueOps {
+		switch op.OpType {
+		case OpTypeSet:
+			err = b.originalStore.SaveValue(op.Key, op.Value)
+		case OpTypeDelete:
+			_, err = b.originalStore.DeleteValue(op.Key)
+		default:
+			err = fmt.Errorf("Invalid Batch operation type: %v", op.OpType)
+		}
+	}
+
+	if err != nil {
+		return
+	}
+
+	for _, op := range b.SegmentOps {
+		switch op.OpType {
+		case OpTypeSet:
+			err = b.originalStore.SaveSegment(op.Segment)
+		case OpTypeDelete:
+			_, err = b.originalStore.DeleteSegment(op.LinkHash)
+		default:
+			err = fmt.Errorf("Invalid Batch operation type: %v", op.OpType)
+		}
+	}
+
+	return
 }
